requirements.txt
fxgui>=2.2.4
colorama>=0.4.6
PyYAML>=6.0.1
tabulate>=0.9.0
pillow>=10.4.0
PySide2>=5.15.0; python_version<"3.11"
PySide6>=6.0.0; python_version>="3.11"
# jupyter
# ipykernel
